ar(feat) add services types
